Types, Hash Table Header, Complex Lexer Test
- defines type aliases for uint8_t => u8, etc.
- uses i32 style types instead of ints, etc.
- defines a header for an hash table implementation
- defines a failing complex lexer test

+++ include/arena.h
@@ -1,22 +1,22 @@
 #ifndef ARENA_H
 #define ARENA_H
 
-#include <stdlib.h>
+#include "types.h"
 
 typedef struct {
 	unsigned char* buf;
-	size_t buf_len;
-	size_t cur_offset;
+	u64 buf_len;
+	u64 cur_offset;
 } Arena;
 
 Arena* arena_init(void);
-void* arena_alloc(Arena* arena, size_t size);
+void* arena_alloc(Arena* arena, u64 size);
 void arena_destroy(Arena* arena);
 
 typedef struct {
 	Arena* arena;
-	size_t cur_offset;
-	size_t prev_offset;
+	u64 cur_offset;
+	u64 prev_offset;
 } Temp_Arena_Memory;
 
 Temp_Arena_Memory temp_area_memory_begin(Arena* arena);
